Update revisedprepdocs.py
For excel, saving the content in the index as a row with the row name(assuming it's in column A) and column names with the data, enabled ChatGPT to respond better.  Not 100% better, i.e. if you have hierarchical data, I found including the parent in the row name helps too.  But that is pre-processing work.